+++ src/errors/errors.ts
@@ -31,8 +31,3 @@ export {
   isNetworkError,
   isTimeoutError,
 } from '@uluops/sdk-core/errors';
-
-export {
-  ResponseValidationError,
-  isResponseValidationError,
-} from '@uluops/sdk-core';